requirements.txt
Flask==2.3.3
SpeechRecognition==3.10.0
pyaudio==0.2.11
openai==1.3.0
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
pydub==0.25.1
numpy==1.24.3
